Rename lamp_S to lamp.sh
Greetings, I’m Md. I’m a consultant living in Dhaka, Bangladesh. I am a fan of Hi-Tech, entrepreneurship, and programming.
I’m also interested in Linux, DevOps, Linux Security, Cisco, Networking and innovation. If you have 
If you need LAMP (Linux, Apache, SQL, PHP) install or any project related with any linux ping me:
       
      Telegram: @zicdhobd
            skype : zicodhkbd
            Gmail : zicodhkbd@gmail.com
               Live : zicodhkbd@hotmail.com
           twitter: zicodhkbd

LAMP stack is an open source web development platform that uses Linux as operating system, Apache as a web server and MySQL or MariaDB as a relational database management system and PHP as the object-oriented scripting language. Sometimes Perl or Python is used instead of PHP in the LAMP lifecycle.

The LAMP stack is the foundation for Linux hosted websites is the Linux, Apache, MySQL and PHP (LAMP) software stack.

LAMP Stand for :-

L (Linex OS)
A (Apache [Local Client Server])
M (MySQL Database)
P (PHP)
A LAMP Stack is a set of open-source software that can be used to create websites and web applications. LAMP is an acronym, and these stacks typically consist of the Linux operating system, the Apache HTTP Server, the MySQL relational database management system, and the PHP programming language.

Linux: The operating system (OS) makes up our first layer. Linux sets the foundation for the stack model. All other layers run on top of this layer. Linux is the lowest-level layer and provides the operating system. Linux actually runs each of the other components. You are not specifically limited to Linux, however; you can easily run each of the other components on Microsoft®; Windows®;, Mac OS X, or UNIX® if you need to.

Apache: The second layer consists of web server software, typically Apache Web Server. This layer resides on top of the Linux layer. Web servers are responsible for translating from web browsers to their correct website. Apache provides the mechanics for getting a Web page to a user. Apache is a stable, mission-critical-capable server, and it runs more than 65 percent of all Web sites on the Internet. The PHP component actually sits inside Apache, and you use Apache and PHP together to create your dynamic pages.

MySQL: Our third layer is where databases live. MySQL stores details that can be queried by scripting to construct a website. MySQL usually sits on top of the Linux layer alongside Apache/layer 2. In high-end configurations, MySQL can be offloaded to a separate host server. MySQL provides the data-storage side of the LAMP system. With MySQL, you have access to a very capable database suitable for running large and complex sites. Within your Web application, all your data, products, accounts, and other types of information will reside in this database in a format that you can easily query with the SQL language.

PHP: Sitting on top of them all is our fourth and final layer. The scripting layer consists of PHP and/or other similar web programming languages. Websites and Web Applications run within this layer. PHP is a simple and efficient programming language that provides the glue for all the other parts of the LAMP system. You use PHP to write dynamic content capable of accessing the data in the MySQL database and some of the features that Linux provides.